+++ lib/convert.ts
@@ -1,22 +1,39 @@
-import { SID_MAP, STATUS_MAP, SYMBOL_MODIFIER_MAP } from "./mappings";
+import {
+  INVERTED_SID_MAP,
+  INVERTED_STATUS_MAP,
+  INVERTED_SYMBOL_MODIFIER_MAP,
+  SID_MAP,
+  STATUS_MAP,
+  SYMBOL_MODIFIER_MAP,
+} from "./mappings";
 
-import letter2number from "./legacydata.json";
+import letter2numberTable from "./legacydata.json";
+import {
+  normalizeLetterCode,
+  parseLetterSidc,
+  parseNumberSidc,
+  replaceCharAt,
+} from "./helpers";
 
-export function replaceCharAt(
-  text: string,
-  index: number,
-  replacementChar: string
-) {
-  return text.substring(0, index) + replacementChar + text.substring(index + 1);
+export interface Letter2NumberOptions {
+  fallbackSidc?: string;
 }
 
-function normalizeLetterCode(sidc: string) {
-  return replaceCharAt(replaceCharAt(sidc, 3, "*"), 1, "*");
+export interface Number2LetterOptions {}
+
+export interface Letter2NumberResult {
+  sidc: string;
+  success: boolean;
+}
+
+export interface Number2LetterResult {
+  sidc: string;
+  success: boolean;
 }
 
 function findSymbol(digits: string): string[] | undefined {
   let beginning = 0,
-    end = letter2number.length,
+    end = letter2numberTable.length,
     target;
   if (!end) {
     return;
@@ -25,35 +42,25 @@ function findSymbol(digits: string): string[] | undefined {
     target = (beginning + end) >> 1;
     if (
       (target === end || target === beginning) &&
-      letter2number[target][0] !== digits
+      letter2numberTable[target][0] !== digits
     ) {
       return;
     }
-    if (letter2number[target][0] > digits) {
+    if (letter2numberTable[target][0] > digits) {
       end = target;
-    } else if (letter2number[target][0] < digits) {
+    } else if (letter2numberTable[target][0] < digits) {
       beginning = target;
     } else {
-      return letter2number[target];
+      return letter2numberTable[target];
     }
   }
-}
-
-export interface Letter2NumberOptions {
-  fallbackSidc?: string;
-}
-
-export interface Letter2NumberResult {
-  sidc: string;
-  success: boolean;
 }
 
 export function convertLetterSidc2NumberSidc(
   letterSidc: string,
   options: Letter2NumberOptions = {}
 ): Letter2NumberResult {
-  const standardIdentity = letterSidc[1];
-  const status = letterSidc[3];
+  const { standardIdentity, status } = parseLetterSidc(letterSidc);
   const symbolModifier = letterSidc.substring(10, 12).replace("*", "-");
 
   const normalizedSidc = normalizeLetterCode(letterSidc).slice(0, 10);
@@ -80,4 +87,37 @@ export function convertLetterCode2NumberCode(
 ): string {
   const { sidc } = convertLetterSidc2NumberSidc(letterSidc, options);
   return sidc;
+}
+
+export function convertNumberSidc2LetterSidc(
+  numberSidc: string,
+  options: Number2LetterOptions = {}
+): Letter2NumberResult {
+  const parts = parseNumberSidc(numberSidc);
+  const status = INVERTED_STATUS_MAP[parts.status];
+  const standardIdentity =
+    INVERTED_SID_MAP[parts.context + parts.standardIdentity];
+
+  const symbolModifier =
+    parts.hqemt === "000" ? "---" : INVERTED_SYMBOL_MODIFIER_MAP[parts.hqemt];
+  const nCode = parts.mainIcon + parts.modifierOne + parts.modifierTwo;
+
+  const match = letter2numberTable.find(
+    ([letterCode, symbolSet, numericCode]) => {
+      return symbolSet === parts.symbolSet && numericCode === nCode;
+    }
+  );
+  let sic = "";
+
+  if (match) {
+    sic = match[0];
+  }
+
+  return {
+    sidc:
+      replaceCharAt(replaceCharAt(sic, 1, standardIdentity), 3, status) +
+      symbolModifier +
+      "---",
+    success: Boolean(match),
+  };
 }